initialize payment

+++ migrations/20200112111126-create_product_table.js
@@ -22,8 +22,8 @@ module.exports = {
       allowNull: false,
     },
     productDiscount: {
-      type: Sequelize.DECIMAL,
-      defaultValue: 0.00,
+      type: Sequelize.INTEGER,
+      defaultValue: 0,
       allowNull: false
     },
     productShipping: {
